Move performance tests to `log4j-perf`
By moving the performance tests from `log4j-core` to `log4j-perf`, we
can drop the `log4j:log4j` dependency
