+++ log4j-core/src/test/java/org/apache/logging/log4j/core/appender/mom/kafka/KafkaAppenderTest.java
@@ -36,8 +36,8 @@ import java.util.concurrent.TimeoutException;
 import org.apache.kafka.clients.producer.MockProducer;
 import org.apache.kafka.clients.producer.ProducerRecord;
 import org.apache.kafka.clients.producer.RecordMetadata;
-import org.apache.log4j.MDC;
 import org.apache.logging.log4j.Level;
+import org.apache.logging.log4j.ThreadContext;
 import org.apache.logging.log4j.categories.Appenders;
 import org.apache.logging.log4j.core.Appender;
 import org.apache.logging.log4j.core.LogEvent;
@@ -61,7 +61,7 @@ public class KafkaAppenderTest {
 
 			Future<RecordMetadata> retVal = super.send(record);
 
-			boolean isRetryTest = "true".equals(MDC.get("KafkaAppenderWithRetryCount"));
+			boolean isRetryTest = "true".equals(ThreadContext.get("KafkaAppenderWithRetryCount"));
 			if (isRetryTest) {
 				try {
 					throw new TimeoutException();
@@ -179,7 +179,7 @@ public class KafkaAppenderTest {
 	@Test
 	public void testAppendWithRetryCount() {
 		try {
-			MDC.put("KafkaAppenderWithRetryCount", "true");
+			ThreadContext.put("KafkaAppenderWithRetryCount", "true");
 			final Appender appender = ctx.getRequiredAppender("KafkaAppenderWithRetryCount");
 			final LogEvent logEvent = createLogEvent();
 			appender.append(logEvent);
@@ -189,7 +189,7 @@ public class KafkaAppenderTest {
 		} catch (Exception e) {
 			e.printStackTrace();
 		} finally {
-			MDC.clear();
+			ThreadContext.clearMap();
 		}
 
 	}
