+++ log4j-perf/src/main/java/org/apache/logging/log4j/perf/async/Histogram.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.logging.log4j.core.async.perftest;
+package org.apache.logging.log4j.perf.async;
 
 import java.math.BigDecimal;
 import java.math.RoundingMode;
