+++ log4j-perf/src/main/java/org/apache/logging/log4j/perf/async/PerfTest.java
@@ -14,7 +14,7 @@
  * See the license for the specific language governing permissions and
  * limitations under the license.
  */
-package org.apache.logging.log4j.core.async.perftest;
+package org.apache.logging.log4j.perf.async;
 
 import java.io.FileWriter;
 import java.io.IOException;
@@ -28,8 +28,8 @@ import org.apache.logging.log4j.core.util.Loader;
  * Single-threaded performance test. Usually invoked from PerfTestDriver as part of a series of tests.
  * <p>
  * To run a single instance of this class for the log4j2 test runner:<br>
- * java -Dlog4j.configurationFile=mylog4j2.xml org.apache.logging.log4j.core.async.perftest.PerfTest \
- * org.apache.logging.log4j.core.async.perftest.RunLog4j2 <name> <resultfile.txt> <-verbose> <-throughput>
+ * java -Dlog4j.configurationFile=mylog4j2.xml org.apache.logging.log4j.perf.async.PerfTest \
+ * org.apache.logging.log4j.perf.async.RunLog4j2 <name> <resultfile.txt> <-verbose> <-throughput>
  */
 public class PerfTest {
