+++ log4j-perf/src/main/java/org/apache/logging/log4j/perf/async/RunLogback.java
@@ -14,7 +14,7 @@
  * See the license for the specific language governing permissions and
  * limitations under the license.
  */
-package org.apache.logging.log4j.core.async.perftest;
+package org.apache.logging.log4j.perf.async;
 
 import org.slf4j.LoggerFactory;
